Move shared OnionRouting codecs to separate file
These utilities can be used outside of payment scenarios, so they deserve
their own file.
